Moving some stuff around to make way for other services using the sql layers.
