+++ packages/backend-core/package.json
@@ -54,7 +54,8 @@
     "sanitize-s3-objectkey": "0.0.1",
     "semver": "^7.5.4",
     "tar-fs": "2.1.1",
-    "uuid": "^8.3.2"
+    "uuid": "^8.3.2",
+    "knex": "2.4.2"
   },
   "devDependencies": {
     "@shopify/jest-koa-mocks": "5.1.1",
