+++ packages/server/src/definitions/datasource.ts
@@ -3,8 +3,3 @@
  * internal to the server and don't need to *
  * be exposed for use by other services.    *
  ********************************************/
-
-export interface QueryOptions {
-  disableReturning?: boolean
-  disableBindings?: boolean
-}
