+++ packages/server/src/integrations/utils/index.ts
@@ -1,2 +1,1 @@
 export * from "./utils"
-export { SqlStatements } from "./sqlStatements"
